Remove: MyTeams 컴포넌트 삭제, 메인페이지로 통합
